use abstract class
