+++ src/LEGO.AsyncAPI.Bindings/Sns/PrincipalObject.cs
@@ -1,17 +1,10 @@
-// Copyright (c) The LEGO Group. All rights reserved.
-
 namespace LEGO.AsyncAPI.Bindings.Sns;
 
 using System;
 using System.Collections.Generic;
 using LEGO.AsyncAPI.Writers;
 
-public interface IPrincipalValue
-{
-    void Serialize(IAsyncApiWriter writer);
-}
-
-public struct PrincipalObject : IPrincipalValue
+public class PrincipalObject : Principal
 {
     private KeyValuePair<string, StringOrStringList> PrincipalValue;
 
@@ -20,7 +13,7 @@ public struct PrincipalObject : IPrincipalValue
         this.PrincipalValue = principalValue;
     }
 
-    public void Serialize(IAsyncApiWriter writer)
+    public override void Serialize(IAsyncApiWriter writer)
     {
         if (writer is null)
         {
@@ -30,25 +23,5 @@ public struct PrincipalObject : IPrincipalValue
         writer.WriteStartObject();
         writer.WriteRequiredObject(this.PrincipalValue.Key, this.PrincipalValue.Value, (w, t) => t.Value.Write(w));
         writer.WriteEndObject();
-    }
-}
-
-public struct PrincipalStar : IPrincipalValue
-{
-    private string PrincipalValue;
-
-    public PrincipalStar()
-    {
-        this.PrincipalValue = "*";
-    }
-
-    public void Serialize(IAsyncApiWriter writer)
-    {
-        if (writer is null)
-        {
-            throw new ArgumentNullException(nameof(writer));
-        }
-
-        writer.WriteValue(this.PrincipalValue);
     }
 }
