+++ src/LEGO.AsyncAPI.Bindings/Sns/Statement.cs
@@ -46,7 +46,7 @@ namespace LEGO.AsyncAPI.Bindings.Sns
 
             writer.WriteStartObject();
             writer.WriteRequiredProperty("effect", this.Effect.GetDisplayName());
-            writer.WriteRequiredObject("principal", this.Principal, (w, t) => t.Value.Serialize(w));
+            writer.WriteRequiredObject("principal", this.Principal, (w, t) => t.Serialize(w));
             writer.WriteRequiredObject("action", this.Action, (w, t) => t.Value.Write(w));
             writer.WriteOptionalObject("resource", this.Resource, (w, t) => t?.Value.Write(w));
             writer.WriteOptionalObject("condition", this.Condition, (w, t) => t?.Write(w));
