+++ src/LEGO.AsyncAPI.Bindings/Sqs/Statement.cs
@@ -47,7 +47,7 @@ namespace LEGO.AsyncAPI.Bindings.Sqs
 
             writer.WriteStartObject();
             writer.WriteRequiredProperty("effect", this.Effect.GetDisplayName());
-            writer.WriteRequiredObject("principal", this.Principal, (w, t) => t.Value.Serialize(w));
+            writer.WriteRequiredObject("principal", this.Principal, (w, t) => t.Serialize(w));
             writer.WriteRequiredObject("action", this.Action, (w, t) => t.Value.Write(w));
             writer.WriteOptionalObject("resource", this.Resource, (w, t) => t?.Value.Write(w));
             writer.WriteOptionalObject("condition", this.Condition, (w, t) => t?.Write(w));
