+++ test/LEGO.AsyncAPI.Tests/Bindings/Sns/SnsBindings_Should.cs
@@ -86,7 +86,7 @@ namespace LEGO.AsyncAPI.Tests.Bindings.Sns
                         new Statement()
                         {
                             Effect = Effect.Deny,
-                            Principal = new Principal(),
+                            Principal = new PrincipalStar(),
                             Action = new StringOrStringList(new AsyncApiAny(new List<string>()
                             {
                                 "sns:Publish",
@@ -105,7 +105,7 @@ namespace LEGO.AsyncAPI.Tests.Bindings.Sns
                         new Statement()
                         {
                             Effect = Effect.Allow,
-                            Principal = new Principal(new KeyValuePair<string, StringOrStringList>(
+                            Principal = new PrincipalObject(new KeyValuePair<string, StringOrStringList>(
                                 "AWS", new StringOrStringList(new AsyncApiAny(new List<string>
                                     { "arn:aws:iam::123456789012:user/alex.wichmann", "arn:aws:iam::123456789012:user/dec.kolakowski" })))),
                             Action = new StringOrStringList(new AsyncApiAny("sns:Create")),
