+++ test/LEGO.AsyncAPI.Tests/Bindings/Sqs/SqsBindings_should.cs
@@ -136,7 +136,7 @@ namespace LEGO.AsyncAPI.Tests.Bindings.Sqs
                             new Statement()
                             {
                                 Effect = Effect.Deny,
-                                Principal = new Principal(new KeyValuePair<string, StringOrStringList>(
+                                Principal = new PrincipalObject(new KeyValuePair<string, StringOrStringList>(
                                     "AWS", new StringOrStringList(new AsyncApiAny("arn:aws:iam::123456789012:user/alex.wichmann")))),
                                 Action = new StringOrStringList(new AsyncApiAny(new List<string>
                                 {
@@ -166,7 +166,7 @@ namespace LEGO.AsyncAPI.Tests.Bindings.Sqs
                             new Statement()
                             {
                                 Effect = Effect.Allow,
-                                Principal = new Principal(new KeyValuePair<string, StringOrStringList>(
+                                Principal = new PrincipalObject(new KeyValuePair<string, StringOrStringList>(
                                     "AWS", new StringOrStringList(new AsyncApiAny(new List<string>
                                         { "arn:aws:iam::123456789012:user/alex.wichmann", "arn:aws:iam::123456789012:user/dec.kolakowski" })))),
                                 Action = new StringOrStringList(new AsyncApiAny("sqs:CreateQueue")),
@@ -223,7 +223,7 @@ namespace LEGO.AsyncAPI.Tests.Bindings.Sqs
                             new Statement()
                             {
                                 Effect = Effect.Allow,
-                                Principal = new Principal(new KeyValuePair<string, StringOrStringList>(
+                                Principal = new PrincipalObject(new KeyValuePair<string, StringOrStringList>(
                                     "Service", new StringOrStringList(new AsyncApiAny("s3.amazonaws.com")))),
                                 Action = new StringOrStringList(new AsyncApiAny(new List<string>
                                 {
@@ -377,7 +377,7 @@ namespace LEGO.AsyncAPI.Tests.Bindings.Sqs
                                 new Statement()
                                 {
                                     Effect = Effect.Deny,
-                                    Principal = new Principal(new KeyValuePair<string, StringOrStringList>(
+                                    Principal = new PrincipalObject(new KeyValuePair<string, StringOrStringList>(
                                         "AWS", new StringOrStringList(new AsyncApiAny("arn:aws:iam::123456789012:user/alex.wichmann")))),
                                     Action = new StringOrStringList(new AsyncApiAny(new List<string>()
                                     {
@@ -398,7 +398,7 @@ namespace LEGO.AsyncAPI.Tests.Bindings.Sqs
                                 new Statement()
                                 {
                                     Effect = Effect.Allow,
-                                    Principal = new Principal(new KeyValuePair<string, StringOrStringList>(
+                                    Principal = new PrincipalObject(new KeyValuePair<string, StringOrStringList>(
                                         "AWS", new StringOrStringList(new AsyncApiAny(new List<string>
                                             { "arn:aws:iam::123456789012:user/alex.wichmann", "arn:aws:iam::123456789012:user/dec.kolakowski" })))),
                                     Action = new StringOrStringList(new AsyncApiAny("sqs:CreateQueue")),
@@ -446,7 +446,7 @@ namespace LEGO.AsyncAPI.Tests.Bindings.Sqs
                                 new Statement()
                                 {
                                     Effect = Effect.Allow,
-                                    Principal = new Principal(new KeyValuePair<string, StringOrStringList>(
+                                    Principal = new PrincipalObject(new KeyValuePair<string, StringOrStringList>(
                                         "AWS", new StringOrStringList(new AsyncApiAny("arn:aws:iam::123456789012:user/alex.wichmann")))),
                                     Action = new StringOrStringList(new AsyncApiAny(new List<string>
                                     {
